requirements.txt
litellm
python-dotenv
rich